feat: Implement More Insights feature with charting and download capabilities
- Added ChartContainer component for displaying insights charts.
- Created ControlsBar for managing chart settings and refreshing data.
- Implemented ErrorFallback component for error handling.
- Added HiddenSitesInfo to inform users about hidden sites.
- Introduced RefreshToast for user feedback on data refresh.
- Developed Sidebar for site selection and actions.
- Established constants for time options and chart types.
- Created custom hooks: useAbortSignal for aborting fetch requests, useDownload for handling data downloads, and useMoreInsights for managing insights state and data fetching.
- Implemented utility functions for CSV formatting and file naming.
- Integrated responsive design with mobile sidebar support.
- Enhanced user experience with loading states and error messages.
