+++ src/platform/src/common/features/download-insights-locations/index.jsx
@@ -6,7 +6,7 @@ import Close from '@/icons/close_icon';
 import PropTypes from 'prop-types';
 import DataDownload, { DownloadDataHeader } from './modules/DataDownload';
 import AddLocations, { AddLocationHeader } from './modules/AddLocations';
-import MoreInsights, { InSightsHeader } from './modules/MoreInsights';
+import MoreInsights, { InSightsHeader } from './more-insights';
 import PlantTree, { AddPlantTreeHeader } from './modules/PlantTree';
 import BuyDevice, { AddBuyDeviceHeader } from './modules/BuyDevice';
 import Search, { AddSearchHeader } from './modules/Search';
