+++ src/platform/src/common/features/download-insights-locations/more-insights/constants/options.js
@@ -1,0 +1,5 @@
+export const TIME_OPTIONS = ['daily', 'hourly', 'monthly'];
+export const CHART_TYPE = [
+  { id: 'line', name: 'Line' },
+  { id: 'bar', name: 'Bar' },
+];
