+++ src/platform/src/common/features/download-insights-locations/more-insights/hooks/useAbortSignal.js
@@ -1,0 +1,11 @@
+import { useRef, useEffect } from 'react';
+
+export default function useAbortSignal() {
+  const abortRef = useRef();
+  useEffect(() => () => abortRef.current?.abort?.(), []);
+  return () => {
+    abortRef.current?.abort?.();
+    abortRef.current = new AbortController();
+    return abortRef.current.signal;
+  };
+}
