+++ src/platform/src/common/features/download-insights-locations/more-insights/utils/mergeAbort.js
@@ -1,0 +1,14 @@
+import { useRef, useEffect } from 'react';
+
+export default function useMergeAbort() {
+  const ref = useRef({});
+  useEffect(
+    () => () => {
+      Object.values(ref.current).forEach((ctrl) =>
+        ctrl?.abort ? ctrl.abort() : clearTimeout(ctrl),
+      );
+    },
+    [],
+  );
+  return ref;
+}
